# booking-sys/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "这是电影院购票系统的后端服务API",
        "title": "Booking System API",
        "termsOfService": "https://github.com/lightsaid/booking-apps",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:5000",
    "basePath": "/v1/api",
    "paths": {
        "/admin/theaters": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "根据page_num和page_size返回电影列表",
                "tags": [
                    "Theaters"
                ],
                "summary": "电影院列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "第几页",
                        "name": "page_num",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "每页多条",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "description": "手机验证码登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "user login param",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.loginUserRequest"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/sms": {
            "post": {
                "description": "发送短信验证码，如果是开发模式会直接返回验证码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Other"
                ],
                "summary": "发送短信验证码",
                "parameters": [
                    {
                        "description": "手机号",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.SMSCode"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.SMSCode": {
            "type": "object",
            "required": [
                "phone_number"
            ],
            "properties": {
                "phone_number": {
                    "type": "string"
                }
            }
        },
        "main.loginUserRequest": {
            "type": "object",
            "required": [
                "code",
                "phone_number"
            ],
            "properties": {
                "code": {
                    "type": "integer",
                    "maximum": 9999,
                    "minimum": 1000
                },
                "phone_number": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}